requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
amqp==2.5.2
appnope==0.1.0
arrow==0.15.6
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
binaryornot==0.4.4
cchardet==2.1.6
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.2
Django==3.0.5
django-extensions==2.2.9
djangorestframework==3.11.0
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
fake-useragent==0.1.11
fancycompleter==0.9.1
feedfinder2==0.0.4
feedparser==5.2.1
fire==0.1.1
geographiclib==1.50
geopy==1.21.0
googletransx==2.4.2
gunicorn==20.0.4
idna==2.5
importlib-metadata==1.6.0
ipython==5.9.0
ipython-genutils==0.2.0
jedi==0.17.0
jieba3k==0.35.1
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.14.1
kombu==4.6.8
lxml==4.5.0
MarkupSafe==1.1.1
multidict==4.7.5
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.3
pandas==1.0.3
parso==0.7.0
pdbpp==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
poyo==0.5.0
prompt-toolkit==1.0.18
psycopg2==2.8.5
ptyprocess==0.6.0
pycares==3.1.1
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyrepl==0.9.0
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.17.3
requests-file==1.5.1
schedule==0.6.0
simplegeneric==0.8.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.45.0
traitlets==4.3.3
twint==2.1.19
urllib3==1.21.1
urlnormalizer==1.2.5
vine==1.3.0
wcwidth==0.1.9
Werkzeug==1.0.1
wmctrl==0.3
yarl==1.4.2
zipp==3.1.0
